checking if merchant exists before attmepting to delete

+++ paypro-android-backend/src/main/java/air/found/payproandroidbackend/business_logic/MerchantService.java
@@ -1,9 +1,12 @@
 package air.found.payproandroidbackend.business_logic;
 
+import air.found.payproandroidbackend.core.models.Merchants;
 import air.found.payproandroidbackend.data_access.MerchantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class MerchantService {
@@ -15,13 +18,13 @@ public class MerchantService {
     }
 
     public boolean deleteMerchantById(Integer merchantId) {
-        try {
+        Optional<Merchants> merchantOptional = merchantsRepository.findById(merchantId);
+
+        if (merchantOptional.isPresent()) {
             merchantsRepository.deleteById(merchantId);
-            return true; // Deletion successful
-        } catch (EmptyResultDataAccessException e) {
-            return false; // Merchant not found
-        } catch (Exception e) {
-            return false; // Error deleting merchant
+            return true;
+        } else {
+            return false;
         }
     }
 }
